functional audio playback

+++ packages/key-finder-web/src/AudioFileKeyDetection/AudioFileItem.tsx
@@ -1,4 +1,4 @@
-import { h, Component } from 'preact';
+import { h, Component, createRef } from 'preact';
 import { keyFinderUtils, audioUtils } from '../Utils';
 import * as fretboards from 'fretboards';
 
@@ -18,6 +18,7 @@ interface Props {
   fileItem: FileItem;
   updateDigest: (uuid: string, digest: string) => void;
   updateResult: (uuid: string, result: string) => void;
+  audioElement: HTMLAudioElement | null; // Pass the audio element from the parent component
 }
 
 interface State {
@@ -32,12 +33,12 @@ interface State {
   scale: { [key: string]: number[] } | null;
   isPlaying: boolean;
   currentTime: number;
-  source: AudioBufferSourceNode | null;
-  playingSource: AudioBufferSourceNode | null;
+  lastPlayedPosition: number | null;
+  isReadyToPlay: boolean;
 }
 
 class AudioFileItem extends Component<Props, State> {
-  audioElement: HTMLAudioElement = new Audio();
+  audioElement: HTMLAudioElement | null = null; // Use HTMLAudioElement for audio playback
   worker: Worker | null = null;
   terminated: boolean = false;
 
@@ -53,115 +54,74 @@ class AudioFileItem extends Component<Props, State> {
     scale: null,
     isPlaying: false,
     currentTime: 0,
-    source: null,
-    playingSource: null,
+    isReadyToPlay: false,
+    lastPlayedPosition: null,
   };
 
   audioContext: AudioContext | null = null;
-  source: AudioBufferSourceNode | null = null;
+  audioSource: AudioBufferSourceNode | null = null;
+  originalSource: AudioBufferSourceNode | null = null;
+  audioInterval: number | null;
+  lastPlayedPosition: number | null;
+  isReadyToPlay: boolean = false;
+  mounted = false;
+  fretboardContainerRef = createRef<HTMLDivElement>();
 
   componentDidMount() {
-    if (this.props.fileItem.canProcess) {
-      const reader = new FileReader();
-      reader.onload = this.handleFileLoad;
-      reader.readAsArrayBuffer(this.props.fileItem.file);
-    }
-
-    // Create the audio context here and use it for playback later
-    this.audioContext = new (window.AudioContext ||
-      (window as any).webkitAudioContext)();
-    this.audioElement.addEventListener('error', (error) => {
-      console.error('Audio playback error:', error);
-    });
+    this.mounted = true; // Set to true when the component is mounted
+    this.initAudio(this.props.fileItem);
   }
 
   componentDidUpdate(prevProps: Props) {
-    if (
-      prevProps.fileItem.canProcess === false &&
-      this.props.fileItem.canProcess === true
-    ) {
-      const reader = new FileReader();
-      reader.onload = this.handleFileLoad;
-      reader.readAsArrayBuffer(this.props.fileItem.file);
+    if (prevProps.fileItem.id !== this.props.fileItem.id) {
+      console.log('AudioFileItem - componentDidUpdate');
+      this.initAudio(this.props.fileItem);
     }
   }
 
   componentWillUnmount() {
+    this.mounted = false; // Set to false when the component is unmounted
     this.worker?.terminate();
-    this.audioElement.pause();
-    this.audioElement.removeEventListener(
+    this.audioElement?.removeEventListener(
+      'canplaythrough',
+      this.handleAudioCanPlay
+    );
+    this.audioElement?.removeEventListener(
       'timeupdate',
       this.handleAudioTimeUpdate
     );
-    this.audioElement.removeEventListener('error', () => {});
-
-    // Stop and disconnect the playingSource if it exists
-    if (this.state.playingSource) {
-      this.state.playingSource.stop();
-      this.state.playingSource.disconnect();
-    }
-
-    // Close the audio context when the component is unmounted
-    if (this.audioContext) {
-      this.audioContext.close();
-      this.audioContext = null;
-    }
   }
 
-  handleAudioPlayPause = () => {
-    if (!this.audioContext) return;
-
-    const { source, isPlaying, currentTime } = this.state;
-
-    if (isPlaying) {
-      this.audioContext.suspend().then(() => {
-        this.setState({ isPlaying: false });
-        source.stop();
-      });
-    } else {
-      this.audioContext.resume().then(() => {
-        if (this.state.playingSource) {
-          // Stop the currently playing source if it exists
-          this.state.playingSource.stop();
-        }
-        this.setState({ isPlaying: true });
-        const newSource = this.audioContext.createBufferSource();
-        newSource.buffer = source.buffer;
-        newSource.connect(this.audioContext.destination);
-        newSource.onended = () => {
-          this.setState({ isPlaying: false, playingSource: null });
-        };
-        this.setState({ playingSource: newSource });
-        newSource.start(0, currentTime); // Start the new source node from the current time
-      });
-    }
+  initAudio = (fileItem) => {
+    // Existing audio setup logic
+    const reader = new FileReader();
+    reader.onload = this.handleFileLoad;
+    reader.readAsArrayBuffer(fileItem.file);
   };
 
-  handleAudioTimeUpdate = () => {
-    this.setState({ currentTime: this.audioElement.currentTime });
-  };
-
-  handleAudioSeek = (event: Event) => {
-    if (!this.audioElement) return;
-    const seekTime = (event.target as HTMLInputElement).valueAsNumber;
-    this.audioElement.currentTime = seekTime;
-    this.setState({ currentTime: seekTime });
+  handleFileLoad = async (event: ProgressEvent<FileReader>): Promise<void> => {
+    const context = audioUtils.createAudioContext();
+    const digest = await crypto.subtle.digest(
+      'SHA-256',
+      event.target.result as ArrayBuffer
+    );
+    const hashArray = Array.from(new Uint8Array(digest));
+    const hashHex = hashArray
+      .map((b) => b.toString(16).padStart(2, '0'))
+      .join('');
+    this.props.updateDigest(this.props.fileItem.id, hashHex);
+    context.decodeAudioData(
+      event.target.result as ArrayBuffer,
+      this.handleAudioFile
+    );
   };
 
   handleAudioFile = async (buffer: AudioBuffer) => {
     console.log('AudioFileItem - handleAudioFile');
 
-    const context = this.audioContext; // Use the existing audio context
-    if (!context) return;
-
-    const source = context.createBufferSource();
-    source.buffer = buffer;
-    source.connect(context.destination);
-    source.onended = () => {
-      this.setState({ isPlaying: false });
-    };
-
-    this.setState({ analyzing: true });
+    this.setState({
+      analyzing: true,
+    });
 
     const worker = keyFinderUtils.initializeKeyFinder({
       sampleRate: buffer.sampleRate,
@@ -182,17 +142,19 @@ class AudioFileItem extends Component<Props, State> {
         console.log(result);
         const normalizedResult = this.getKeySignatureNumericValue(result);
         console.log(normalizedResult);
-        this.setState(
-          {
-            result,
-            analyzing: false,
-          },
-          () => {
-            this.props.updateResult(this.props.fileItem.id, result);
-            worker.terminate();
-            this.renderFretboardScale(normalizedResult); // Render the fretboard scale
-          }
-        );
+        if (this.mounted) {
+          this.setState(
+            {
+              result,
+              analyzing: false,
+            },
+            () => {
+              this.props.updateResult(this.props.fileItem.id, result);
+              worker.terminate();
+              this.renderFretboardScale(normalizedResult); // Render the fretboard scale
+            }
+          );
+        }
       } else {
         // Not final response
         if (event.data.data === 0) {
@@ -235,11 +197,38 @@ class AudioFileItem extends Component<Props, State> {
       buffer.numberOfChannels,
       0
     );
-
-    // Start audio playback after the context is created and everything is set up
-    source.start();
-    this.setState({ source, isPlaying: true });
   };
+
+  getKeySignatureNumericValue(result: string | null) {
+    console.log('AudioFileItem - getKeySignatureNumericValue');
+    if (!result) {
+      return null;
+    }
+
+    const normalizedResult = result.toLowerCase();
+
+    return normalizedResult;
+  }
+
+  arrayBufferToHex(buffer: ArrayBuffer): string {
+    const hashArray = Array.from(new Uint8Array(buffer));
+    return hashArray.map((b) => b.toString(16).padStart(2, '0')).join('');
+  }
+
+  renderFretboardScale(normalizedResult: string | null) {
+    if (!normalizedResult) {
+      return;
+    }
+
+    const container = this.fretboardContainerRef.current;
+
+    if (container) {
+      container.innerHTML = '';
+
+      const fb = fretboards.Fretboard();
+      fb.add(normalizedResult).paint(container);
+    }
+  }
 
   advanceSegmentCount = (
     worker: Worker,
@@ -261,98 +250,101 @@ class AudioFileItem extends Component<Props, State> {
     }));
   };
 
-  getKeySignatureNumericValue(result: string | null) {
-    console.log('AudioFileItem - getKeySignatureNumericValue');
-    if (!result) {
-      return null;
-    }
+  handleAudioCanPlay = () => {
+    console.log('AudioFileItem - handleAudioCanPlay');
+    this.setState({ analyzing: false });
 
-    const normalizedResult = result.toLowerCase();
+    const audioElement = this.props.audioElement; // Use the audioElement from props
 
-    return normalizedResult;
-  }
+    if (!audioElement) return;
 
-  handleFileLoad = async (event: ProgressEvent<FileReader>): Promise<void> => {
-    const context = new (window.AudioContext ||
-      (window as any).webkitAudioContext)();
-    if (!crypto?.subtle?.digest) {
-      console.error(
-        'Web Cryptography API is not supported in this environment.'
-      );
+    if (!audioElement.paused && !audioElement.ended) {
+      // Only play the audio if it's not already playing and not ended
       return;
     }
-    const digest = await crypto.subtle.digest(
-      'SHA-256',
-      event.target.result as ArrayBuffer
-    );
-    const hashArray = Array.from(new Uint8Array(digest));
-    const hashHex = hashArray
-      .map((b) => b.toString(16).padStart(2, '0'))
-      .join('');
-    this.props.updateDigest(this.props.fileItem.id, hashHex);
 
-    console.log('AudioFileItem - handleFileLoad');
-    context.decodeAudioData(
-      event.target.result as ArrayBuffer,
-      (buffer) => {
-        this.handleAudioFile(buffer);
-      },
-      (error) => {
-        console.error('Error decoding audio data:', error);
-      }
-    );
+    audioElement.currentTime = this.state.lastPlayedPosition || 0;
+    if (this.state.isPlaying) {
+      audioElement.play();
+    }
   };
-  renderFretboardScale(normalizedResult: string | null) {
-    if (!normalizedResult) {
-      return;
+
+  handleAudioPlayPause = () => {
+    const audioElement = this.props.audioElement; // Use the audioElement from props
+    if (!audioElement) return;
+
+    if (audioElement.paused || audioElement.ended) {
+      audioElement.play();
+    } else {
+      audioElement.pause();
     }
 
-    const container = document.getElementById('fretboard-container');
+    this.setState((prevState) => ({ isPlaying: !prevState.isPlaying }));
+  };
 
-    // Clear the container before rendering the fretboard to avoid duplicates
-    container.innerHTML = '';
+  handleAudioSeek = (event: Event) => {
+    const audioElement = this.props.audioElement; // Use the audioElement from props
+    if (!audioElement || !Number.isFinite(audioElement.duration)) return;
 
-    const fb = fretboards.Fretboard();
-    fb.add(normalizedResult).paint(container);
-  }
+    const seekTime = (event.target as HTMLInputElement).valueAsNumber;
+
+    // Make sure the seekTime is within the valid range (0 to duration)
+    const validSeekTime = Math.max(
+      0,
+      Math.min(seekTime, audioElement.duration)
+    );
+
+    // Check if the validSeekTime is finite before updating the currentTime
+    if (Number.isFinite(validSeekTime)) {
+      // Update the currentTime on the audioElement
+      audioElement.currentTime = validSeekTime;
+    }
+  };
+
+  handleAudioTimeUpdate = () => {
+    const audioElement = this.props.audioElement; // Use the audioElement from props
+    if (!audioElement) return;
+
+    this.setState({ currentTime: audioElement.currentTime });
+  };
 
   render() {
     const { fileItem } = this.props;
-    const { analyzing, result, isPlaying, currentTime } = this.state;
+    const { analyzing, result } = this.state;
 
     return (
       <div class="file-item__container">
-        <div class="file-item__info">{/* Render file item information */}</div>
+        {/* ... (existing code) , fix innerhtml not rendering 7.20*/}
         <div class="file-item__rendered-fretboard">
           {analyzing && <div>Analyzing...</div>}
           {result && (
             <div>
               Result: {result}
-              {/* Render the fretboard scale here */}
-              <div id="fretboard-container"></div>
+              {/* ... (existing code) */}
             </div>
           )}
+
+          <div ref={this.fretboardContainerRef} />
 
           {/* Audio Player */}
           <div>
             {fileItem.file && (
               <audio
-                ref={(el) => (this.audioElement = el)}
                 src={URL.createObjectURL(fileItem.file)}
-                controls // Add the controls attribute here
-                onTimeUpdate={this.handleAudioTimeUpdate}
+                // ref={(el) => (this.audioElement = el)}
+                controls={true} // Disable default controls
               />
             )}
-            <button onClick={this.handleAudioPlayPause}>
-              {isPlaying ? 'Pause' : 'Play'}
+            <button onClick={() => this.handleAudioPlayPause()}>
+              {this.props.audioElement?.paused ? 'Play' : 'Pause'}
             </button>
             {fileItem.file && (
               <input
                 type="range"
                 min={0}
-                max={this.audioElement?.duration || 0}
-                value={currentTime}
-                onChange={this.handleAudioSeek}
+                max={this.props.audioElement?.duration || 0}
+                value={this.props.audioElement?.currentTime || 0}
+                onInput={this.handleAudioSeek}
               />
             )}
           </div>
